requirements.txt
python-telegram-bot==20.6
PyMuPDF==1.23.7
